feat: Add user authorization info to /api/user/settings (#1034)
