+++ api/config.go
@@ -40,27 +40,6 @@ type Config struct {
 	Authorization Authorization
 }
 
-// Authorization contains authorization configuration.
-type Authorization struct {
-	AdminList           map[string]struct{}
-	Enabled             bool
-	AllowedContactTypes map[string]struct{}
-}
-
-// IsEnabled returns true if auth is enabled and false otherwise.
-func (auth *Authorization) IsEnabled() bool {
-	return auth.Enabled
-}
-
-// IsAdmin checks whether given user is considered an administrator.
-func (auth *Authorization) IsAdmin(login string) bool {
-	if !auth.IsEnabled() {
-		return false
-	}
-	_, ok := auth.AdminList[login]
-	return ok
-}
-
 // WebConfig is container for web ui configuration parameters.
 type WebConfig struct {
 	SupportEmail         string                `json:"supportEmail,omitempty" example:"opensource@skbkontur.com"`
