+++ api/controller/user.go
@@ -12,7 +12,7 @@ func GetUserSettings(database moira.Database, userLogin string, auth *api.Author
 		User: dto.User{
 			Login:       userLogin,
 			AuthEnabled: auth.IsEnabled(),
-			Role:        dto.GetRole(userLogin, auth),
+			Role:        auth.GetRole(userLogin),
 		},
 		Contacts:      make([]moira.ContactData, 0),
 		Subscriptions: make([]moira.SubscriptionData, 0),
